channeldb: Add keychain data and functions
This adds two new buckets and a new function to the channel database:
buckets for tracking keychain data and key family indexes and a function
to return unique indices for individual key families.

The main purpose of this is to allow keyring implementations to store
the used indices for their keys without having to rely on an external
store. This is mostly useful for using wallet-backed root HD keys while
the wallet itself does not track off-chain keys and addresses.

This change is self-contained and does not require database migrations.