sweep: wrap weight estimator
Preparation for a cpfp-aware weight estimator. For cpfp, a regular
weight estimator isn't sufficient, because it needs to take into account
the weight of the parent transaction(s) as well.

Note(decred): the file new file is not renamed and is left with its
content intact in this commit to make it easier to track changes. It
will be fixed in the following commit.