breacharbiter_test: assert publication of adjusted justice tx
Since the breacharbiter will publish an adjusted justice tx after
detecing a spend from the breached commitment, we could get into a race
where sometime it would not get the second spend before publication,
resulting in a deadlock on the publTx channel.

Now we instead wait for the publication of this adjusted tx before
notifying the second spend.