lnwallet: Only perform account discovery once
This adds a new record in the database that tracks whether account
discovery is needed on the wallet. By default, dcrlnd will perform an
account discovery process and will only mark as account discovery not
needed once the wallet is fully synced at least once.

This is only applicable to the embedded wallet, given that remote
wallets are controlled externally.