rpcserver: Report startup router pruning progress
This adds reporting about the initial graph pruning performed by the
channel router. This is usually the longest startup stage of ln wallets
that have been offline for longer periods of time, so reporting it at
the standard log level (info) and via the GetInfo grpc call improves the
UX for startup.