dcrlnd: Start initchainsync service after unlocking
This makes the start up procedure of the app offer a new RPC service in
the time between when the wallet is created or unlocked and the full set
of RPC services are provided.

This new service can be used to track the progress of the initial chain
sync process and is useful for clients that connect to a dcrlnd instance
that is still in the initial startup stage to provide better feedback to
the user.