peer: Stop writeHandler after stalls
This makes the writeHandler of a peer exit with an error whenever the
time between sends is larger than the maximum idle timeout of the
connection, which causes a local disconnection.

This is needed to handle situations where the computer running dcrlnd is
suspended and then resumes after enough time has passed for the remote
peer to have disconnected.

The current version of the time package in the Go stdlib does not
account for time suspended when handling network deadlines and timers.
This causes sends across the connection to silently fail after the
computer resumes. Depending on specific timings and operations started
after resuming, this can cause payments to enter a "limbo" state, where
they are not correctly handled by dcrlnd.

This alleviates the issue by preemptively closing the peer connection
whenever a stall is detected just before writing a new message over the
wire, which causes upstream subsystems to not enter into erroneous
states.

While technically the issue still remains, because the suspend could
happen at any time after the check is performed but before the bytes are
actually flushed across the network, this interval is small enough to
significantly improve the current situation and empirical testing with
this commit applied does not reproduce the reported issue.