multi: Update go.mod to v1.18
This updates the required go version in go.mod to go 1.18. Consequently,
it fixes all formatting errors generated with golangci-lint to adhere to
the new requirements.

Additionally, it updates the required version of the
github.com/stretchr/testify module to version 1.8.0 and updates the
AES module to github.com/Yawning/aez to drop the need for a replace
directive.