multi: Update dcrd and dcrwallet to their master version
This updates the dcrd and dcrwallet dependencies to their current
master wip versions. This is done to ensure continued compatibility of
the master dcrlnd branch to the master dcrd and dcrwallet branches.

The only notable API changes were adding a context argument to rpctest
and switching the CheckTransactionSanity function from the blockchain to
the standalone package.