multi: Switch to dcrtest package
This switches instances of the old dcrd/rpctest package to the recently
released dcrtest/dcrdtest package. In the future, the dcrd/rpctest
package will be removed.

Additionally, it switches instances of the simnet miner to use the one
provided by the dcrdtest package and removes the local implementation of
the miner.