chainscan: Fix issues when when adding targets during historical scan
This fixes some issues when adding new targets using the callback during
a scan.

The underlying issue is that the endHeight for newly added targets was
not set, causing the target to immediately become stale and be removed
from the historical search if it was added via the callback.

This would cause the search to terminate early, potentially missing
events for the target.

The solution added in this commit is to track which targets were added
during a particular call to scan(), and then process these targets the
same way that targets added via Find() were, thus ensuring that their
endHeight would be set if it was originally unset.

To allow this, the calls to the callback now enforce that the
startHeight of new targets is necessarily the current block height + 1.
The documentation was updated to point out this fact and a test to
assert the correct behavior is added.