routing: Handle known spent edges
This adds a check and early return for processing channel announcements
for edges that are known to be spent (due to their corresponding UTXO
having been spent on-chain).

This improves performance of SPV wallets by avoiding a costly rescan
during the GetUtxo calls when processing announcements for channels that
have already been closed.