routing: Refresh annoucements only for open channels
This introduces a change to the router to only consider open channels
when iterating through the list of local node edges for purposes of
refreshing channel announcements.

This is meant to address a (so far) unidentified bug that causes a node
to see one of its already closed channels in its local graph database.
This causes the node to keep refreshing the channel, even when it cannot
possibly be used for routing anymore. This ends up preventing downstream
nodes from dropping the channel from their own graph databases.

While the implementation is less than ideal performance-wise (due to
requiring loading the entire OpenChannel structure), this is only called
once every 24h, therefore it's not on a hot code path.