etcd: Make it build
This makes the etcd kvdb driver build and pass its unit tests. This will
make it easier to bring in future upstream work.

Note that due to walletdb having been updated to v1.4.0 (as opposed to
the upstream version of v1.3.5 at this point in the upstream history),
some additional changes are needed. A port of the upstream commit
fc139d9bb from PR 5245 is also added.