main: Disable anchor commitments in mainnet
Some features of anchor commitments are not usable in the current
Decred's chain, therefore we disable them in mainnet. They are
maintained in other networks in order to allow testing.

Note that due to these limitations, some integration tests need to be
skipped. Specifically, those that rely on the HTLC outputs being
aggregated into a single sweeping transaction.