lnwallet: Fix flaky spend unconfirmed test
This fixes a flaky test that involves first spending the entire wallet
balance to then spend an unconfirmed utxo.

Due to specific timings, the actual utxos in the wallet could be
different than a fixed number, therefore we modify to calculate the
required fee dynamically, based on all the utxos actually in Bob's
wallet.