channeldb/migration: copy current lnwire to migration dir
To avoid code changing underneath the static migrations, we copy the
lnwire dependency in its current form into the migration directory.

Ideally the migrations doesn't depend on any code that might change,
this takes us a step closer.

Note(decred): This is a port of the upstream commit 7569cca19, written
individually instead of cherry-picked.