channeldb/migration21: copy over legacy types+codec for migration21
The legacy encoding depends on the lnwire21 version of lnwire, so it
will let us change lnwire after the migration. To make sure it is
separated from the new encoding, we add it to a new package 'legacy'.

We also put common types in a new package 'common', which will house
types that won't change during the migration, and can be used by both
legacy and current serialization code.

Note(decred): This is a direct port of the upstream commit 4133b4d04.