testutils: Use IPC and random port for dcrwallet instances
This makes the dcrwallet initialized for tests use a random port to bind
its gRPC service, and then use the wallet provided IPC subsystem to read
the gRPC port back.

This is more reliable than trying to determine an open port for the
wallet to use and prevents test flakes when multiple remote wallets are
started.