itest: Cleanup pending HTLCs in force closed channel
This ensures the cleanupForceClose assertion attempts to cleanup any
pending HTLCs in the target channel when cleaning up a force-closed
channel.

This ensures any HTLCs not yet asserted to have been swept are given an
attempt to do so.

This will be necessary to ensure future tests won't leave the channel in
a pending state.