breacharbiter: Consume initial epoch notification before revenge
This drains the initial notification for the current block height from
the epoch notifier in the breach arbiter's revenge loop before listening
for new block notifications.

This is needed to avoid prematurely broadcasting the multiple justice
transactions variants of the revenge logic when the node was restarted
with multiple blocks mined. The node will require determining HTLC
outputs that have already been spent to a second-level transaction,
before any of the justice transactions have a chance to succeed and
acting on the first block notification means it would act _immediately_
upon the restart.

This fixes the dcrlnd-only test case
testRevokedCloseRetributionRemoteHodlSecondLevel.