itest: Assert topology change after opening channel
This is a port of the upstream commit b78164. It ensures that clients
see not just the channel itself in their list of channels, but also that
the routing subsystem registers the topology change, such that the
channels are suitable for routing.

It also adds an additional check to the
testUpdateChannelPolicyForPrivateChannel to attempt to prevent it from
flaking.