multi: Track ChannelReestablish wait time
This adds tracking for how long a peer has been online, while NOT
reestasblishing a channel through the use of ChannelReestablish P2P
messages.

If a peer has been online for a long time without sending a
ChannelReestablish, this is a hint that perhaps their end of the channel
is borked and cannot be used anymore.  One situation where this occurs
is when the peer restores its node without use of an SCB file, in which
case the ChannelReestablish messages the local peer sends are ignored
and the remote peer does not reply with their own ChannelReestablish.