multi: Automatically close non-reestablished channels
This adds automation for closing channels that haven't been
reestablished for some time using the recently introduced metric.

Channels with peers that have been online for some time but which have
NOT been reestablished though the use of ChannelReestablish messages are
good candidates for being force closed, because it is likely the remote
peer has lost the ability to use them.

Given that the metric tracks the total time across restarts and only
while the remote peer is online, we use a default of 72h for the
threshold to auto close the channel.  This should be a reasonable
compromise between not closing too fast on hubs (that are online 24/7)
and ephemeral nodes (that may be online only for an hour or two a day).

A config parameter is added to control the threshold time used for
autoclosing and an itest is added that asserts the correct behavior.