chainntnfs: Delay spend/confirm notifications
Several subsystems (in particular, the gossiper) rely on having updated
their "current block height" field _before_ a spend or confirm
notification happens.

However, due to the asyncness nature of notification dispatching the
current design for the notifications cannot ensure that reliably happens
when the machine has high load.

One concrete manifestation of this issue is on CI runs for itests that
use the dcrd notification implementations, where the confirm
notifications arrives before the block notification, causing a flake.

This fixes the issue by adding an artificial interval between block and
spend/confirm notifications on the notifiers.