breacharbiter: Wait for all spend events before recreating justice tx
The breach arbiter waits for spend events of inputs of a recently
created justice transaction and, after processing them, creates a new
justice tx for any missing inputs.

When the tx notifier subsystem is slow in notifying about the spends,
the function that watches for the spend events may have only partially
processed the actually spent inputs, therefore erroneously triggering
the creation of another justice tx.

This can be manifested on SPV wallets by the presence of a new justice
tx that double spends some of the breach inputs, and as a flake during
the revoked_uncooperative_close_retribution_remote_hodl itest (in the
SPV backend).

To fix this issue, an arbitrary delay of 10 milliseconds is added to
ensure all spend notifications have been received by the processing
goroutines before a new justice transaction is attempted. In the
aforementioned test, this ensures that all inputs have been swept
already and that the final balance in the wallet is correct with no
extraneous unconfirmed transactions.