channeldb: Add dcrlnd migration 01
This migration fixes the prior migration 20, introduced when porting the
upstream code.

That migration erroneously encoded outpoints without the Tree field that
exists in decred code, thus rendering the index incorrect.

The new migration corrects the issue by assuming the tree of every entry
is zero, which is true because the channels can only reside in regular
(as opposed to stake) transactions.