chainntfns: Prevent double sending of Update ntfns
TxNotifier's ConnectTip() and NotifyHeight() are not atomic. Therefore,
the ending height used for a historical search may still be notified for
as a match for a tip confirmation when NotifyHeight() is called.

When using the chainscan driver, the timings are such that registering
for a conf notification just as a block with the target is coming in may
race, and produce double write to the Updates chan, causing a deadlock.

This is evidenced by the historical_conf_dispatch_with_script_dispatch
test of the chainntfns interface test when running with the
chainscan-based driver.

This error is not triggered in upstream lnd because the mining,
notification and historical search intervals are faster.

This is fixed by this commit by double checking whether the full conf
event has already been dispatched or if there are any pending updates to
be read by the client of the conf notification before sending one that
will deadlock.