peer: Use a dedicated ping logger
This switches the peer to log ping/pong messages using a separate logger
with a dedicated subsystem tag. This is useful when running in debug
mode for extended periods of time, as ping messages are generally
useless but repeat constantly, needlessly growing the size of log files
and making debug harder.