channeldb: Remove gossip timestamps in the future
This adds a migration that removes gossip timestamps that are in the
future from the database of last received peer gossip timestamp.

This fixes issues that could lead to lack of routing information when
the peer wrongly sent a date in the future as the timestamp for a
channel update or node announcement.