discovery: Handle future gossip timestamps
This adds some handling for gossip timestamps in the future.

Previously, a wrongly configured node could send channel updates or node
announcements with a timestamp in the future, and the propagation of
such timestamps would cause the local node to record a time in the
future as the most recent timestamp received via gossip.

This would cause that, on restarts, the node would install a gossip
timestamp horizon far in the future, causing it to miss channel updates
and node announcements and breaking pathfinding.

This fixes the immediate handling of future timestamps in the gossiper,
while a future commit will fix the existing database entries with a
migration.

The fix in the gossiper involves disregarding future timestamps for the
purposes of storing it in the db, and (for future proofing against
this error happening again) not using such timestamps in case they are
actually saved in the db.