chainview: Simplify reorg interface test
This removes the initial part of the testFilterBlockDisconnected, which
causes a reorg all the way to genesis, in favor of only testing the
smaller reorg of an already setup chain.

The reorg to genesis will never actually happen in mainnet and SPV
wallets have some trouble dealing with large reorgs all the way to
genesis (specifically: the peers may get dropped due to too many
orphan/sidechain headers being sent).

In practice, decred reorgs are limited to a depth of 6 blocks due to PoS
validation, thus exercising a 10 block reorg is sufficient for testing
purposes.